fix(omo-senpi): coalesce omo.task.updated snapshot pushes to stop RPC queue overflow bursts
Every accepted child progress event of every live task emitted a full
session snapshot, so a burst of distinct ticks pushed hundreds of snapshots
back-to-back and tore down the desktop RPC connection (senpi#1438; the 4 MiB
consumer cap is intentional and unchanged). Progress-driven pushes now run
through a 150ms trailing-edge window, latest-wins: the first tick after idle
schedules one flush, later ticks only mark state dirty, and the flush emits
a single snapshot of the freshest store reads. Fingerprint dedupe and the
256-task cap are retained, attach()/sync() still emit immediately and cancel
any pending flush so ordering stays monotonic, and detach()/dispose() cancel
pending timers so nothing emits after teardown. Timers are injectable the
same way as the status UI debounce; the production timer is unref'd.
