+++ packages/omo-senpi/src/components/task/AGENTS.md
@@ -15,7 +15,7 @@ Senpi adapter for the `@oh-my-opencode/senpi-task` engine: task/team tools, the 
 | engine node ops (`packages/senpi-task/src/dag/`) | `node-retry.ts` gives failed/cancelled/skipped nodes a fresh attempt and un-skips their cascaded dependents; `node-send.ts` steers a running node's child or revives a finished resident one; `node-control-context.ts` is the shared resolution/validation seam both use. Node-scoped errors: `node_not_found`, `node_not_retryable`, `node_not_continuable`, `amend_running_node`, `invalid_amendment`, `run_still_active`. |
 | `dag-rpc-bridge.ts` / `dag-rpc-handlers.ts` | Three wire channels: sequenced `omo.dag.event` ledger, unsequenced telemetry, snapshot push; four query handlers sharing one parser vocabulary and error envelope. Contract types in `dag-rpc-bridge-contract.ts`, payloads in `dag-snapshot-payload.ts`. |
 | `dag-wake.ts` / `dag-wake-source.ts` | Idle-injection wake messages (`omo-senpi.dag-run`) and read-only `wake_source_state` liveness under source `omo-dag`. |
-| `task-rpc-bridge.ts` / `task-rpc-codec.ts` | `omo.task.updated` snapshot push plus task send/cancel/output over RPC; the codec bounds every field (`st_` id pattern, 32k text caps, 1000 tail lines). |
+| `task-rpc-bridge.ts` / `task-rpc-codec.ts` | `omo.task.updated` snapshot push plus task send/cancel/output over RPC; the codec bounds every field (`st_` id pattern, 32k text caps, 1000 tail lines). Progress-driven pushes coalesce on a 150ms trailing edge (latest-wins, injectable `TaskRpcTimers`) so a burst of child ticks cannot overflow the consumer socket queue; `attach`/`sync` still emit immediately and cancel any pending flush. |
 | `team-service.ts` / `team-service-support.ts` | Lead-side team CRUD, mailbox reconcile, shutdown request/approve/reject over `team-core` storage; spec normalization + member validation in the support module. The team lead is ALWAYS the current session (no model-supplied override); run ids are canonical UUID-v4, and foreign owned-team run ids are rejected, never answered with empty data. |
 | `lead-poller-lifecycle.ts` | One 1-second lead poller per team led by the current session; delivery journal + idle-injection sink. |
 | `member-liveness.ts` / `owned-member-liveness.ts` | Death notifications for owned members (`senpi-task.team-member-liveness`); suspended residency states (`persisted_only`/`rpc_detached`) are skipped, not dead. |
