+++ packages/omo-senpi/src/components/task/task-rpc-bridge.ts
@@ -22,6 +22,13 @@ import {
 
 const TASK_UPDATED_EVENT = "omo.task.updated"
 const MAX_TASK_SNAPSHOTS = 256
+// Trailing-edge window for progress-driven snapshot pushes. Every accepted child progress event of
+// every live task used to emit a full session snapshot (up to 256 records), so a burst of distinct
+// ticks pushed hundreds of snapshots back-to-back and overflowed the desktop RPC socket queue
+// (senpi#1438; the 4 MiB cap and teardown-on-overflow on the consumer side are intentional). 150ms
+// is below a UI flicker threshold yet collapses a whole burst into one push of the freshest state;
+// lifecycle-driven attach/sync still emit immediately, so only progress tail latency is affected.
+const PROGRESS_COALESCE_MS = 150
 
 type TimerHandle = ReturnType<typeof setTimeout> | number
 
@@ -59,11 +66,19 @@ export function wireTaskRpcBridge(
   deps: TaskRpcBridgeDeps = {},
 ): TaskRpcBridge {
   const timers = deps.timers ?? globalTimers
+  const coalesceMs = deps.coalesceMs ?? PROGRESS_COALESCE_MS
   const subscriptions = new Map<string, () => void>()
   const liveProgress = new Map<string, TaskLiveProgressSnapshot>()
   let activeSessionId: string | undefined
   let lastSnapshot: string | undefined
   let disposed = false
+  let pendingFlush: TimerHandle | undefined
+
+  const cancelPendingFlush = (): void => {
+    if (pendingFlush === undefined) return
+    timers.clear(pendingFlush)
+    pendingFlush = undefined
+  }
 
   const recordsForSession = (sessionId: string) => {
     const all = engine.manager.list({ scope: "parent-session", session_id: sessionId }).map((entry) => entry.record)
@@ -101,6 +116,17 @@ export function wireTaskRpcBridge(
     pi.rpc.emit(TASK_UPDATED_EVENT, data)
   }
 
+  const scheduleProgressFlush = (): void => {
+    if (disposed || activeSessionId === undefined) return
+    // Trailing edge, latest-wins: the first tick opens the window, later ticks only mean the
+    // flush re-reads the store, so a burst costs exactly one snapshot push.
+    if (pendingFlush !== undefined) return
+    pendingFlush = timers.set(() => {
+      pendingFlush = undefined
+      emit()
+    }, coalesceMs)
+  }
+
   const removeSubscription = (taskId: string): void => {
     subscriptions.get(taskId)?.()
     subscriptions.delete(taskId)
@@ -134,13 +160,14 @@ export function wireTaskRpcBridge(
       const unsubscribe = engine.manager.subscribeChild(record.task_id, (event: ManagedChildEvent) => {
         if (!progress.accept(event)) return
         liveProgress.set(record.task_id, liveProgressSnapshot(progress.details()))
-        emit()
+        scheduleProgressFlush()
       })
       subscriptions.set(record.task_id, unsubscribe)
     }
   }
 
   const detach = (): void => {
+    cancelPendingFlush()
     for (const taskId of [...subscriptions.keys()]) removeSubscription(taskId)
     activeSessionId = undefined
     lastSnapshot = undefined
@@ -149,6 +176,9 @@ export function wireTaskRpcBridge(
   const sync = (): void => {
     const sessionId = activeSessionId
     if (sessionId === undefined || pi.rpc?.emit === undefined) return
+    // Emit immediately and drop any queued progress flush so ordering stays monotonic: the
+    // lifecycle snapshot below is at least as fresh as anything the flush would have sent.
+    cancelPendingFlush()
     const selection = recordsForSession(sessionId)
     syncSubscriptions(selection.records)
     emit(selection)
